Rename files to follow naming pattern
