--- libs/core/src/metrics/NetworkConditionReason/index.ts
@@ -1,1 +1,0 @@
-export { default, networkConditionReasonValue } from './NetworkConditionReason';
