+++ libs/core/src/metrics/NetworkConditionReasonValue/NetworkConditionReasonValue.ts
